server/requirements.txt
asgiref==3.4.1
Django==3.2.18
django-cors-headers==3.10.1
django-environ==0.9.0
djangorestframework==3.14.0
gunicorn==20.1.0
mysqlclient==2.1.1
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.1.1
